Moved the Human.java in the src directory
